fixed failing test
